+++ libsquoosh/src/image_data.ts
@@ -1,13 +1,9 @@
 export default class ImageData {
-  readonly data: Uint8ClampedArray | Uint8Array;
+  readonly data: Uint8ClampedArray;
   readonly width: number;
   readonly height: number;
 
-  constructor(
-    data: Uint8ClampedArray | Uint8Array,
-    width: number,
-    height: number,
-  ) {
+  constructor(data: Uint8ClampedArray, width: number, height: number) {
     this.data = data;
     this.width = width;
     this.height = height;
